Extract Pane from class
move to listening framework
